+++ gazelle/bzl/testdata/defaultvisibility/BUILD.out
