+++ gazelle/bzl/testdata/defaultvisibility/foo.bzl
