+++ gazelle/bzl/testdata/defaultvisibility/nested/dir/BUILD.in
