+++ gazelle/bzl/testdata/defaultvisibility/nested/dir/BUILD.out
