+++ gazelle/bzl/testdata/defaultvisibility/nested/dir/bar.bzl
