+++ gazelle/bzl/testdata/empty/BUILD.in
