+++ gazelle/bzl/testdata/empty/BUILD.out
