+++ gazelle/bzl/testdata/external/BUILD.in
