+++ gazelle/bzl/testdata/external/BUILD.out
