+++ gazelle/bzl/testdata/external/foo.bzl
