+++ gazelle/bzl/testdata/import/bar.bzl
