+++ gazelle/bzl/testdata/import/foo.bzl
