+++ gazelle/bzl/testdata/multidir/BUILD.out
