+++ gazelle/bzl/testdata/multidir/foo.bzl
