+++ gazelle/bzl/testdata/multidir/nested/dir/BUILD.out
