+++ gazelle/bzl/testdata/multidir/nested/dir/bar.bzl
