+++ gazelle/bzl/testdata/nobuildfiles/BUILD.out
