+++ gazelle/bzl/testdata/private/foo.bzl
