+++ gazelle/bzl/testdata/private/private/bar.bzl
