+++ gazelle/bzl/testdata/simple/BUILD.in
