+++ gazelle/bzl/testdata/simple/BUILD.out
